+++ pyqcrbox/pyqcrbox/data_management/__init__.py
@@ -1,4 +1,15 @@
-from .data_file import Dataset, DatasetInfoResponse
-from .data_file_manager import CalculationAlreadyExists, DataFileManager, DatasetNotFoundError
-from .dummy_data_file_manager import DummyDataFileManager
-from .nats_data_file_manager import NatsDataFileManager
+from .data_file import DataFile, DataFileResponse
+from .data_manager import CalculationAlreadyExists, DataManager, DatasetNotFoundError
+from .dataset import Dataset, DatasetResponse
+from .nats_data_manager import NatsDataManager
+
+__all__ = [
+    "DataFile",
+    "DataFileResponse",
+    "CalculationAlreadyExists",
+    "DataManager",
+    "DatasetNotFoundError",
+    "Dataset",
+    "DatasetResponse",
+    "NatsDataManager",
+]