+++ pyqcrbox/pyqcrbox/registry/server/api/__init__.py
@@ -1,1 +1,5 @@
 from .api_endpoints import api_router
+
+__all__ = [
+    "api_router",
+]